requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.2.1
werkzeug==2.3.7
email-validator==2.0.0
pytz==2023.3
python-dotenv==1.0.0
flask-mail==0.9.1
itsdangerous==2.1.2
flask-bcrypt==1.0.1

# Required for SSL/HTTPS (needed for camera access)
pyopenssl==23.2.0

# Google Gemini AI integration
google-generativeai>=0.3.0

# Security packages
bleach==6.1.0
flask-talisman==1.1.0
wtforms-validators==1.0.0
marshmallow==3.20.1
pydantic==2.5.3
html-sanitizer==2.2.0
flask-limiter==3.5.0
gunicorn==23.0.0